server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
docker
PyGithub
requests
python-dotenv
supabase
github3.py